+++ R/cloc-package.r
@@ -54,6 +54,7 @@
 #' @name cloc-package
 #' @docType package
 #' @author Bob Rudis (bob@@rud.is)
+#' @importFrom git2r clone
 #' @importFrom utils read.table contrib.url download.file download.packages tail
 #' @importFrom dplyr bind_rows progress_estimated data_frame
 #' @importFrom processx run